moved the generation of the SVG sprite directly in the Ember addon “update” function
